commit3:overflow solved!

+++ project1.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    int a = 0, b = 0;
-    scanf("%d %d", &a, &b);
-    printf("%d", a * b);
+    long int a = 0, b = 0;
+    scanf("%ld %ld", &a, &b);
+    printf("%ld", a * b);
 }
